Move to first batch of final Polymon art assets.
Includes tweaks to the carousel and general layout of the start screen.
Also notably moves away from using an SVG sprite sheet for all art
assets. This kind of sucks, but the final monster art is too dense and
complicated to ship as raw SVGs, and previous testing suggests that
there are hidden downsides to a giant PNG sprite sheet.